.claude/skills/card-news/scripts/requirements.txt
playwright
google-genai
rembg
Pillow
python-dotenv
